+++ src/test/kotlin/kotools/types/string/NotBlankStringTest.kt
@@ -337,6 +337,92 @@ class NotBlankStringTest {
     }
 
     @Nested
+    inner class ToPositiveInt {
+        @Test
+        fun `should pass with a string representation of 0`() {
+            // GIVEN
+            val value = 0
+            val string = NotBlankString("$value")
+            // WHEN
+            val result: PositiveInt = assertPass(string::toPositiveInt)
+            // THEN
+            result.value assertEquals value
+        }
+
+        @Test
+        fun `should pass with a string representation of 1`() {
+            // GIVEN
+            val value = 1
+            val string = NotBlankString("$value")
+            // WHEN
+            val result: PositiveInt = assertPass(string::toPositiveInt)
+            // THEN
+            result.value assertEquals value
+        }
+
+        @Test
+        fun `should throw an error with an invalid string`() {
+            // GIVEN
+            val string = NotBlankString("a")
+            // WHEN & THEN
+            assertFailsWith<NumberFormatException>(string::toPositiveInt)
+        }
+
+        @Test
+        fun `should throw an error with a string representation of -1`() {
+            // GIVEN
+            val string = NotBlankString("-1")
+            // WHEN & THEN
+            assertFailsWith<IllegalArgumentException>(string::toPositiveInt)
+        }
+    }
+
+    @Nested
+    inner class ToPositiveIntOrNull {
+        @Test
+        fun `should pass with a string representation of 0`() {
+            // GIVEN
+            val value = 0
+            val string = NotBlankString("$value")
+            // WHEN
+            val result: PositiveInt? = string.toPositiveIntOrNull()
+            // THEN
+            result.assertNotNull().value assertEquals value
+        }
+
+        @Test
+        fun `should pass with a string representation of 1`() {
+            // GIVEN
+            val value = 1
+            val string = NotBlankString("$value")
+            // WHEN
+            val result: PositiveInt? = string.toPositiveIntOrNull()
+            // THEN
+            result.assertNotNull().value assertEquals value
+        }
+
+        @Test
+        fun `should return null with an invalid string`() {
+            // GIVEN
+            val string = NotBlankString("a")
+            // WHEN
+            val result: PositiveInt? = string.toPositiveIntOrNull()
+            // THEN
+            result.assertNull()
+        }
+
+        @Test
+        fun `should return null with a string representation of -1`() {
+            // GIVEN
+            val string = NotBlankString("-1")
+            // WHEN
+            val result: PositiveInt? = string.toPositiveIntOrNull()
+            // THEN
+            result.assertNull()
+        }
+    }
+
+    @Nested
     inner class ToString {
         @Test
         fun `should return its value`() {