+++ changelog.md
@@ -19,8 +19,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Conversions from `String` and `NotBlankString` to `NonZeroInt`
-  ([#48](https://github.com/kotools/types/issues/48)).
+- Conversions from `String` and `NotBlankString` types to `NonZeroInt` and
+  `PositiveInt` ([#48](https://github.com/kotools/types/issues/48)).
 
 ## Version 2.0.0 - 2022/08/01